Refactor mmv_image code
